rename lookup, put Instance type disable on edit
