+++ awx/ui/src/components/Lookup/index.js
@@ -8,4 +8,4 @@ export { default as ApplicationLookup } from './ApplicationLookup';
 export { default as HostFilterLookup } from './HostFilterLookup';
 export { default as OrganizationLookup } from './OrganizationLookup';
 export { default as ExecutionEnvironmentLookup } from './ExecutionEnvironmentLookup';
-export { default as InstancesLookup } from './InstancesLookup';
+export { default as PeersLookup } from './PeersLookup';
